Remove syntax::common module
