+++ lang/elaborator/src/typechecker/exprs/local_comatch.rs
@@ -8,7 +8,6 @@ use crate::typechecker::exprs::CheckTelescope;
 use crate::unifier::unify::*;
 use miette_util::ToMiette;
 use syntax::ast::*;
-use syntax::common::*;
 use syntax::ctx::LevelCtx;
 
 use super::super::ctx::*;
