+++ lang/lowering/src/ctx.rs
@@ -10,7 +10,7 @@ use syntax::ast::lookup_table::DeclMeta;
 use syntax::ast::{self, MetaVar, MetaVarState};
 use syntax::ast::{HasSpan, Named};
 use syntax::ast::{HashMap, HashSet};
-use syntax::common::*;
+use syntax::ast::{Idx, Lvl};
 use syntax::ctx::{Context, ContextElem, LevelCtx};
 
 use super::result::LoweringError;
