+++ lang/syntax/src/ast/traits/subst.rs
@@ -2,7 +2,6 @@ use std::rc::Rc;
 
 use crate::ast::Variable;
 use crate::ast::*;
-use crate::common::*;
 use crate::ctx::*;
 
 // Substitution
