--- lang/syntax/src/common/mod.rs
@@ -1,3 +1,0 @@
-mod de_bruijn;
-
-pub use de_bruijn::*;
