+++ lang/syntax/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod ast;
-pub mod common;
 pub mod ctx;
